Extract storybook module
